extract app-selector component
